resources-js | recompile with simple js
